roll out simplified ionchannels
