+++ tests/test_ionchannels.py
@@ -4,10 +4,10 @@ import sympy as sp
 
 from neat.channels.channelcollection import channelcollection
 from neat import IonChannel
-from neat import IonChannelSimplified
 
 import pytest
 import pickle
+import os, shutil
 
 
 class TestChannels():
@@ -15,6 +15,12 @@ class TestChannels():
         tcn = channelcollection.TestChannel()
         v_arr = np.linspace(-80., -10., 10)
 
+        factors = np.array([5.,1.])
+        powers = np.array([[3,3,1],
+                           [2,2,1]])
+
+        varnames = np.array([[sp.symbols('a00'), sp.symbols('a01'), sp.symbols('a02')],
+                             [sp.symbols('a10'), sp.symbols('a11'), sp.symbols('a12')]])
         # state variable asymptotic values
         def varinf(v):
             aux = np.ones_like(v) if isinstance(v, np.ndarray) else 1.
@@ -36,109 +42,113 @@ class TestChannels():
 
         # test whether activations are correct
         var_inf = varinf(v_arr)
-        for ind, f_vi in np.ndenumerate(tcn.f_varinf):
-            var_inf_ind = f_vi(v_arr)
-            assert np.allclose(var_inf_ind, var_inf[ind])
+        var_inf_chan = tcn.computeVarinf(v_arr)
+        for ind, varname in np.ndenumerate(varnames):
+            assert np.allclose(var_inf[ind], var_inf_chan[varname])
+
         # test whether open probability is correct
-        p_open = np.sum(tcn.factors[:, np.newaxis] * \
-                        np.product(var_inf ** tcn.powers[:, :, np.newaxis], 1), 0)
+        p_open = np.sum(factors[:, np.newaxis] * \
+                        np.product(var_inf ** powers[:, :, np.newaxis], 1), 0)
         p_open_ = tcn.computePOpen(v_arr)
         assert np.allclose(p_open_, p_open)
+
         # test whether derivatives are correct
-        dp_dx_arr, df_dv_arr, df_dx_arr = tcn.computeDerivatives(v_arr)
+        dp_dx_chan, df_dv_chan, df_dx_chan = tcn.computeDerivatives(v_arr)
+
         # first: derivatives of open probability
-        for ind, name in np.ndenumerate(tcn.varnames):
-            dp_dx = tcn.factors[ind[0]] * tcn.powers[ind] * \
-                    np.prod(var_inf[ind[0]] ** tcn.powers[ind[0]][:, np.newaxis], 0) / var_inf[ind]
-            assert np.allclose(dp_dx_arr[ind], dp_dx)
+        for ind, varname in np.ndenumerate(varnames):
+            dp_dx = factors[ind[0]] * powers[ind] * \
+                    np.prod(var_inf[ind[0]] ** powers[ind[0]][:, np.newaxis], 0) / var_inf[ind]
+            assert np.allclose(dp_dx_chan[varname], dp_dx)
+
         # second: derivatives of state variable functions to voltage
         df_dv = dvarinf_dv(v_arr) / taurel(v_arr)
-        for ind, name in np.ndenumerate(tcn.varnames):
-            assert np.allclose(df_dv[ind], df_dv_arr[ind])
+        for ind, varname in np.ndenumerate(varnames):
+            assert np.allclose(df_dv[ind], df_dv_chan[varname])
+
         # third: derivatives of state variable functions to state variables
         df_dx = -1. / taurel(v_arr)
-        for ind, name in np.ndenumerate(tcn.varnames):
-            assert np.allclose(df_dx[ind], df_dx_arr[ind])
-        # test whether sympy expressions are correct
-        # TODO
-
-class TestNa(IonChannel):
-    def __init__(self):
-        ## USER DEFINED
-        # ion the ion channel current uses (don't define for unspecific)
-        ion = 'na'
-        # define open probability
-        p_open = sp.sympify('m**3 * h')
-        # define state variable activations and timescales
-        spalpham = sp.sympify('   0.182 * (v + 38.) / (1. - exp(-(v + 38.) / 6.))', evaluate=False) # 1/ms
-        spbetam  = sp.sympify(' - 0.124 * (v + 38.) / (1. - exp( (v + 38.) / 6.))', evaluate=False) # 1/ms
-        spalphah = sp.sympify(' - 0.015 * (v + 66.) / (1. - exp( (v + 66.) / 6.))', evaluate=False) # 1/ms
-        spbetah  = sp.sympify('   0.015 * (v + 66.) / (1. - exp(-(v + 66.) / 6.))', evaluate=False) # 1/ms
-        m_inf = spalpham / (spalpham + spbetam)
-        h_inf = spalphah / (spalphah + spbetah)
-        tau_m = (1./2.95) / (spalpham + spbetam)
-        tau_h = (1./2.95) / (spalphah + spbetah)
+        for ind, varname in np.ndenumerate(varnames):
+            assert np.allclose(df_dx[ind], df_dx_chan[varname])
 
 
-        ## INTERNAL
-        # this array fixes the ordering of symbols
-        # --> PROBLEM: the order of the symbols is not fixed, hence this might
-        # cause confusion whenever a lambdified function is called
-        self.statevars = np.array([symbol for symbol in p_open.free_symbols])
-        # redundant, but current IonChannel implementation uses it
-        self.varnames = np.array([str(sv) for sv in self.statevars])
-
-        # define symbols
-        for symbol in p_open.free_symbols:
-            exec(str(symbol) + ' = symbol')
-
-        # store open probability as attribute
-        self.p_open = p_open
-
-        # store v sympy symbol
-        self.sp_v = [symb for symb in m_inf.free_symbols][0]
-        print('---', self.sp_v)
-
-        # define state variable function arrays
-        # we have to maintain the same ordering of statevariables to be able to
-        # keep track of lambda function arguments
-        self.fstatevar = np.zeros_like(self.statevars)
-        self.varinf = np.zeros_like(self.statevars)
-        self.tauinf = np.zeros_like(self.statevars)
-        for ind, var in np.ndenumerate(self.statevars):
-            self.varinf[ind] = eval(str(var)+'_inf')
-            self.tauinf[ind] = eval('tau_' +str(var))
-            self.fstatevar[ind] = eval('('+str(var)+'_inf - '+ str(var) +') / tau_'+str(var))
-
-        # concentrations, for if the ion channel uses it
-        if not hasattr(self, 'ion'):
-            self.ion = ''
-        if not hasattr(self, 'concentrations'):
-            self.concentrations = []
-        self.sp_c = [sp.symbols(conc) for conc in self.concentrations]
-
-        # set lambda functions
-        self.setLambdaFuncs()
-
-    def testPOpen(self):
-        na = channelcollection.Na_Ta()
-
-        p_o_1 = na.computePOpen(-35.)
-        p_o_2 = self.computePOpen(-35.)
-        assert np.allclose(p_o_1, p_o_2)
-
-    def testLinSum(self):
-        na = channelcollection.Na_Ta()
-
-        l_s_1 = na.computeLinSum(-35., 0., 50.)
-        l_s_2 = self.computeLinSum(-35., 0., 50.)
-        assert np.allclose(l_s_1, l_s_2)
+# class TestNa(IonChannel):
+#     def __init__(self):
+#         ## USER DEFINED
+#         # ion the ion channel current uses (don't define for unspecific)
+#         ion = 'na'
+#         # define open probability
+#         p_open = sp.sympify('m**3 * h')
+#         # define state variable activations and timescales
+#         spalpham = sp.sympify('   0.182 * (v + 38.) / (1. - exp(-(v + 38.) / 6.))', evaluate=False) # 1/ms
+#         spbetam  = sp.sympify(' - 0.124 * (v + 38.) / (1. - exp( (v + 38.) / 6.))', evaluate=False) # 1/ms
+#         spalphah = sp.sympify(' - 0.015 * (v + 66.) / (1. - exp( (v + 66.) / 6.))', evaluate=False) # 1/ms
+#         spbetah  = sp.sympify('   0.015 * (v + 66.) / (1. - exp(-(v + 66.) / 6.))', evaluate=False) # 1/ms
+#         m_inf = spalpham / (spalpham + spbetam)
+#         h_inf = spalphah / (spalphah + spbetah)
+#         tau_m = (1./2.95) / (spalpham + spbetam)
+#         tau_h = (1./2.95) / (spalphah + spbetah)
 
 
-def test_na():
-    tna = TestNa()
-    tna.testPOpen()
-    tna.testLinSum()
+#         ## INTERNAL
+#         # this array fixes the ordering of symbols
+#         # --> PROBLEM: the order of the symbols is not fixed, hence this might
+#         # cause confusion whenever a lambdified function is called
+#         self.statevars = np.array([symbol for symbol in p_open.free_symbols])
+#         # redundant, but current IonChannel implementation uses it
+#         self.varnames = np.array([str(sv) for sv in self.statevars])
+
+#         # define symbols
+#         for symbol in p_open.free_symbols:
+#             exec(str(symbol) + ' = symbol')
+
+#         # store open probability as attribute
+#         self.p_open = p_open
+
+#         # store v sympy symbol
+#         self.sp_v = [symb for symb in m_inf.free_symbols][0]
+#         print('---', self.sp_v)
+
+#         # define state variable function arrays
+#         # we have to maintain the same ordering of statevariables to be able to
+#         # keep track of lambda function arguments
+#         self.fstatevar = np.zeros_like(self.statevars)
+#         self.varinf = np.zeros_like(self.statevars)
+#         self.tauinf = np.zeros_like(self.statevars)
+#         for ind, var in np.ndenumerate(self.statevars):
+#             self.varinf[ind] = eval(str(var)+'_inf')
+#             self.tauinf[ind] = eval('tau_' +str(var))
+#             self.fstatevar[ind] = eval('('+str(var)+'_inf - '+ str(var) +') / tau_'+str(var))
+
+#         # concentrations, for if the ion channel uses it
+#         if not hasattr(self, 'ion'):
+#             self.ion = ''
+#         if not hasattr(self, 'concentrations'):
+#             self.concentrations = []
+#         self.sp_c = [sp.symbols(conc) for conc in self.concentrations]
+
+#         # set lambda functions
+#         self._lambdifyChannel()
+
+#     def testPOpen(self):
+#         na = channelcollection.Na_Ta()
+
+#         p_o_1 = na.computePOpen(-35.)
+#         p_o_2 = self.computePOpen(-35.)
+#         assert np.allclose(p_o_1, p_o_2)
+
+#     def testLinSum(self):
+#         na = channelcollection.Na_Ta()
+
+#         l_s_1 = na.computeLinSum(-35., 0., 50.)
+#         l_s_2 = self.computeLinSum(-35., 0., 50.)
+#         assert np.allclose(l_s_1, l_s_2)
+
+
+# def test_na():
+#     tna = TestNa()
+#     tna.testPOpen()
+#     tna.testLinSum()
 
 
 
@@ -147,7 +157,7 @@ def sp_exp(x):
     return sp.exp(x, evaluate=False)
 
 
-class Na_Ta_new(IonChannelSimplified):
+class Na_Ta_new(IonChannel):
     def define(self):
         """
         (Colbert and Pan, 2002)
@@ -190,22 +200,25 @@ class Na_Ta_new(IonChannelSimplified):
         print(exprs)
 
 
-def test_ionchannel_simplified():
+def test_ionchannel_simplified(remove=True):
+    if not os.path.exists('mech/'):
+        os.mkdir('mech/')
+
     na = channelcollection.Na_Ta()
-    na_simplified = channelcollection.Na_Ta_simplified()
 
-    p_o_1 = na.computePOpen(-35.)
-    p_o_2 = na_simplified.computePOpen(-35.)
-    assert np.allclose(p_o_1, p_o_2)
+    p_o = na.computePOpen(-35.)
+    assert np.allclose(p_o, 0.002009216860105564)
 
-    l_s_1 = na.computeLinSum(-35., 0., 50.)
-    l_s_2 = na_simplified.computeLinSum(-35., 0., 50.)
-    assert np.allclose(l_s_1, l_s_2)
+    l_s = na.computeLinSum(-35., 0., 50.)
+    assert np.allclose(l_s, -0.00534261017220376)
 
-    na_simplified.writeModFile('mech/')
+    na.writeModFile('mech/')
 
-    sk = channelcollection.SK_simplified()
+    sk = channelcollection.SK()
     sk.writeModFile('mech/')
+
+    if remove:
+        shutil.rmtree('mech/')
 
 
 def test_pickling():
@@ -223,14 +236,14 @@ def test_pickling():
 
 
 if __name__ == '__main__':
-    # tcns = TestChannels()
-    # tcns.testBasic()
+    tcns = TestChannels()
+    tcns.testBasic()
 
     # tna = TestNa()
     # tna.testPOpen()
     # tna.testLinSum()
 
-    test_ionchannel_simplified()
+    # test_ionchannel_simplified()
 
     # Na_Ta_new()
